pulumi/requirements.txt
pulumi>=3.0.0
pulumi-gcp>=7.0.0
python-dotenv>=1.0.0
